gulpfile has to be part of generator
